chore(acp): split payload tree and text views
